Sensor AppOp check: checkOp -> noteOp
'checkOp' delivers the status of the Op without updating the stats, whilst
noteOp also updates the statistics.

Up to Android 11, the code using 'checkOp' also has at least 'published'
the Op status, so it could e.g. be queried with the shell appops command
or caused the app to be shown, as intended, in the special permissions
dialogue for sensors.

In Android 12L, this seems different, hence the 'noteOp' method is now
called instead to make the functionality work as intended.

Change-Id: I690aa3cd0ad4bead0c0b3524c54178c42fc7af4f